server/requirements.txt
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
channels==1.1.5
constantly==15.1.0
daphne==1.3.0
Django==1.11.2
django-cors-headers==2.1.0
ecdsa==0.13
future==0.16.0
hyperlink==17.2.1
incremental==17.5.0
jws==0.1.3
pycrypto==2.6.1
PyJWT==1.5.0
python-jose==1.3.2
pytz==2017.2
six==1.10.0
Twisted==17.5.0
txaio==2.8.0
zope.interface==4.4.2
